pyproject.toml
[project]
name = "ai-model-dev"
version = "0.2.0"
description = "Personal Research Portal for LLM jailbreaking safety research"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40.0",
    "chromadb>=1.0.0",
    "docling>=2.71.0",
    "fpdf2>=2.8.0",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.41.0",
]
